+++ sdk/tables/azure-data-tables/src/main/java/com/azure/data/tables/BuilderHelper.java
@@ -26,10 +26,18 @@ import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.data.tables.implementation.CosmosPatchTransformPolicy;
 import com.azure.data.tables.implementation.NullHttpClient;
+import com.azure.data.tables.implementation.StorageAuthenticationSettings;
+import com.azure.data.tables.implementation.StorageConnectionString;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.StringJoiner;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static com.azure.data.tables.implementation.TableUtils.sasTokenEquals;
 
 final class BuilderHelper {
     private static final Map<String, String> PROPERTIES =
@@ -81,6 +89,7 @@ final class BuilderHelper {
         policies.add(retryPolicy);
 
         policies.add(new AddDatePolicy());
+
         HttpPipelinePolicy credentialPolicy;
 
         if (azureNamedKeyCredential != null) {
@@ -90,16 +99,12 @@ final class BuilderHelper {
         } else if (sasToken != null) {
             credentialPolicy = new AzureSasCredentialPolicy(new AzureSasCredential(sasToken), false);
         } else {
-            credentialPolicy = null;
-        }
-
-        if (credentialPolicy != null) {
-            policies.add(credentialPolicy);
-        } else {
             throw logger.logExceptionAsError(
                 new IllegalStateException("A form of authentication is required to create a client. Use a builder's "
                     + "'credential()', 'sasToken()' or 'connectionString()' methods to set a form of authentication."));
         }
+
+        policies.add(credentialPolicy);
 
         // Add per retry additional policies.
         policies.addAll(perRetryAdditionalPolicies);
@@ -123,5 +128,58 @@ final class BuilderHelper {
             .policies(policies)
             .httpClient(new NullHttpClient())
             .build();
+    }
+
+    static void validateCredentials(AzureNamedKeyCredential azureNamedKeyCredential,
+                                    AzureSasCredential azureSasCredential, String sasToken, String connectionString,
+                                    ClientLogger logger) {
+        List<Object> usedCredentials =
+            Stream.of(azureNamedKeyCredential, azureSasCredential, sasToken, connectionString)
+                .filter(Objects::nonNull)
+                .collect(Collectors.toList());
+
+        // Only allow two forms of authentication when 'connectionString' and 'sasToken' are provided. Validate that
+        // both contain the same SAS settings.
+        if (usedCredentials.size() == 2 && connectionString != null && sasToken != null) {
+            StorageConnectionString storageConnectionString =
+                StorageConnectionString.create(connectionString, logger);
+            StorageAuthenticationSettings authSettings = storageConnectionString.getStorageAuthSettings();
+
+            if (authSettings.getType() == StorageAuthenticationSettings.Type.SAS_TOKEN) {
+                if (sasTokenEquals(sasToken, authSettings.getSasToken())) {
+                    return;
+                } else {
+                    throw logger.logExceptionAsError(new IllegalStateException("'connectionString' contains a SAS token"
+                        + " with different settings than 'sasToken'."));
+                }
+            }
+
+            // If the 'connectionString' auth type is not SAS_TOKEN and a 'sasToken' was provided, then multiplte
+            // incompatible forms of authentication were specified in the client builder.
+        }
+
+        if (usedCredentials.size() > 1) {
+            StringJoiner usedCredentialsStringBuilder = new StringJoiner(", ");
+
+            if (azureNamedKeyCredential != null) {
+                usedCredentialsStringBuilder.add("azureNamedKeyCredential");
+            }
+
+            if (azureSasCredential != null) {
+                usedCredentialsStringBuilder.add("azureSasCredential");
+            }
+
+            if (sasToken != null) {
+                usedCredentialsStringBuilder.add("sasToken");
+            }
+
+            if (connectionString != null) {
+                usedCredentialsStringBuilder.add("connectionString");
+            }
+
+            throw logger.logExceptionAsError(new IllegalStateException(
+                "Only one form of authentication should be used. The authentication forms present are: "
+                    + usedCredentialsStringBuilder + "."));
+        }
     }
 }